+++ examples/with-next-wallet-button/CHANGELOG.md
@@ -1,5 +1,13 @@
 # with-next-wallet-button
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [725a376]
+- Updated dependencies [9be5452]
+  - @rainbow-me/rainbowkit@2.1.1
+
 ## 0.0.18
 
 ### Patch Changes